Final upload of garage C project
This project is intended to incorporate many of the C programming features learned in
this course for application in a practical “real-life” situation. Your task is to construct a
working program which uses a linked list of structures to manage 8 fields.
Information to keep track of for each driver:Name of Mechanic
Name of Garage
Car Info (Year, Make, Model, owner)
Automotive Problem
Part required
Cost of Part
Labor (in hours)
Cost of repair (calculated)
Use linked lists to store the structures and incorporate all the features listed below:
(1) Consider the user of your program at all times when prompting for a response from
the keyboard and how the information will be displayed to the user.
(2) When you start your program, the user should be presented with 2 options:
Start with new data or load previously stored data.
(3) Options to integrate: Add new entries
Browse entries
Modify entry
Delete entry
Save data
Sort into alphabetical order (for display - scrolling)
Search for a Person’s name
Tally up the month’s earnings from all the repairs (labor)
Strategy:
Start small (3 repairs or less). Add each of the options one at a time and be sure to
retest your program before adding the next option. You are to pass in :
- All source code (final version)
- A file database which will load your last “save”.